chore: cleanup pages
